Made result display behave consistantly
Previously if the displayed output was a square root, it would be displayed on the left hand side of the screen, but if it was not a square root, it would display on the right hand side of the screen.

This has now been fixed and both types of result will display on the left hand side of the screen.